Split skcuda pipeline into different file
- Avoids saturating GPU memory (because of no garbage collection)
